+++ packages/moonstone/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 The following is a curated list of changes in the Enact moonstone module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `moonstone/Scroller` not to scroll to wrong position via 5way navigation in RTL languages
+- `moonstone/Slider` to forward `onActivate` event
+- `moonstone/GridListImageItem` to properly vertically align when the content varies in size
+
 ## [2.0.1] - 2018-08-01
 
 ### Fixed